:label:rename: 패키지 정리 및 리팩토링
